+++ Assets/Scripts/Projectile/Projectile.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class Projectile : MonoBehaviour
 {
 	[SerializeField]
-	private int damage = 1;
+	private float damage = 1;
 
 	private void OnTriggerEnter2D(Collider2D collision)
 	{
